+++ tests/models/edx/test_browser.py
@@ -8,9 +8,7 @@ from hypothesis import given, provisional, settings
 from hypothesis import strategies as st
 from pydantic.error_wrappers import ValidationError
 
-from ralph.exceptions import UnknownEventException
-from ralph.models.edx.browser import BaseBrowserEvent, PageClose
-from ralph.models.selector import ModelSelector
+from ralph.models.edx.browser import BaseBrowserEvent
 
 
 @settings(max_examples=1)
@@ -44,19 +42,3 @@ def test_models_edx_browser_base_browser_event_with_invalid_content(
 
     with pytest.raises(ValidationError, match=error):
         BaseBrowserEvent(**invalid_event)
-
-
-@settings(max_examples=1)
-@given(st.builds(PageClose, referer=provisional.urls(), page=provisional.urls()))
-def test_models_selector_model_selector_get_model_with_valid_event(event):
-    """Tests given a page_close event the get_model method should return PageClose model."""
-
-    event = json.loads(event.json())
-    assert ModelSelector(module="ralph.models.edx").get_model(event) is PageClose
-
-
-def test_models_selector_model_selector_get_model_with_invalid_event():
-    """Tests given a page_close event the get_model method should raise UnknownEventException."""
-
-    with pytest.raises(UnknownEventException):
-        ModelSelector(module="ralph.models.edx").get_model({"invalid": "event"})